refactor(api): move player creation to clubs/{clubId}/players and apply ClubAdmin policy
